+++ src/tr2/game/console/common.c
@@ -1,6 +1,0 @@
-#include "game/console/common.h"
-
-void Console_DrawBackdrop(void)
-{
-    // TODO: implement me
-}